+++ src/FormBuilder.ts
@@ -2,4 +2,4 @@
  * Created by Samuel Gratzl on 08.03.2017.
  */
 
-export {FormBuilder, FormElementType, IFormElementDesc, IFormInputTextDesc, IFormInputTextElement, IFormSelect2, IFormSelectOption, IFormSelectElement, IFormSelectDesc, IFormSelectOptions} from './form';
+export {FormBuilder, FormElementType, IFormElementDesc, IFormInputTextDesc, IFormSelect2, IFormSelectOption, IFormSelectDesc, IFormSelectOptions} from './form';
